resolved warnings

+++ Source/Core/FeatureDetection/CornerHarrisDetector.cpp
@@ -22,7 +22,7 @@ std::vector<FeatureContainer> CornerHarrisDetector::StartDetection(std::vector<I
     std::vector<cv::Mat> newImage;
     std::vector<FeatureContainer> featureContainerList;
 
-    int numImages = imageContainerList.size();
+    uint32_t numImages = static_cast<uint32_t>(imageContainerList.size());
     int imgIdx = 1;
 
     // Detector parameters
